@@ -43,11 +43,11 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
 
-        <pf4j.version>3.8.0</pf4j.version>
-        <spring.version>5.2.9.RELEASE</spring.version>
-        <slf4j.version>1.7.25</slf4j.version>
+        <pf4j.version>3.13.0</pf4j.version>
+        <spring.version>6.2.2</spring.version>
+        <slf4j.version>2.0.16</slf4j.version>
 
         <junit.version>4.12</junit.version>
         <mockito.version>2.0.28-beta</mockito.version>
